refactor: extract trial-run constants into shared analysis/trial_run.py
DRY up the duplicated MIN_RUNS_PER_FUZZER, MIN_BUDGET_HOURS, and
warning text into a single analysis/trial_run.py module imported by
benchmark_report.py, invariant_overlap_report.py, and
generate_docs_site.py.
