+++ analysis/benchmark_report.py
@@ -7,6 +7,10 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Dict, List, Optional
 
+_REPO_ROOT = Path(__file__).resolve().parents[1]
+if str(_REPO_ROOT) not in sys.path:
+    sys.path.insert(0, str(_REPO_ROOT))
+
 import matplotlib
 
 matplotlib.use("Agg")
@@ -14,17 +18,9 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 
-REQUIRED_COLS = ["fuzzer", "run_id", "time_hours", "bugs_found"]
+from analysis.trial_run import format_trial_run_warning, is_trial_run
 
-MIN_RUNS_PER_FUZZER = 10
-MIN_BUDGET_HOURS = 24.0
-TRIAL_RUN_WARNING = (
-    "**Warning — trial run.** "
-    "This benchmark was executed with fewer than {n} instances per fuzzer and/or "
-    "a time budget shorter than {t}h. "
-    "Results from trial runs are meant for debugging purposes and are "
-    "not valid for extracting conclusions across different fuzzers."
-)
+REQUIRED_COLS = ["fuzzer", "run_id", "time_hours", "bugs_found"]
 REQUIRED_SAMPLE_BASE_COLS = ["fuzzer", "run_id", "instance_id", "elapsed_seconds"]
 THROUGHPUT_SAMPLE_VALUE_COLS = ["tx_per_second", "gas_per_second"]
 PROGRESS_SAMPLE_VALUE_COLS = [
@@ -38,21 +34,6 @@ PROGRESS_SAMPLE_VALUE_COLS = [
 
 def die(msg: str) -> None:
     raise SystemExit(f"error: {msg}")
-
-
-def is_trial_run(budget: float, runs_per_fuzzer: List[int]) -> bool:
-    if budget < MIN_BUDGET_HOURS:
-        return True
-    if runs_per_fuzzer and min(runs_per_fuzzer) < MIN_RUNS_PER_FUZZER:
-        return True
-    return False
-
-
-def append_trial_run_warning(lines: List[str]) -> None:
-    lines.append(
-        "> " + TRIAL_RUN_WARNING.format(n=MIN_RUNS_PER_FUZZER, t=int(MIN_BUDGET_HOURS))
-    )
-    lines.append("")
 
 
 def load_csv(path: Path) -> pd.DataFrame:
@@ -999,7 +980,8 @@ def write_report(
     lines.append(f"- Time budget: **{budget:.2f}h**")
     lines.append("")
     if is_trial_run(budget, [m.runs for m in metrics]):
-        append_trial_run_warning(lines)
+        lines.append("> " + format_trial_run_warning())
+        lines.append("")
     lines.append("## Executive summary")
     lines.append(
         "This report is derived solely from cumulative bugs-found over time across repeated runs per fuzzer. "
@@ -1116,7 +1098,8 @@ def write_no_data_report(
     lines.append(f"- Source CSV: `{csv_path}`")
     lines.append("")
     if is_trial_run(budget, []):
-        append_trial_run_warning(lines)
+        lines.append("> " + format_trial_run_warning())
+        lines.append("")
     lines.append("## No data")
     lines.append("")
     lines.append(
